requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
gunicorn==23.0.0
importlib-metadata==6.7.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.7.1
whitenoise==6.5.0
zipp==3.15.0
